requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.19.0
pyserial>=3.5.0
pyzbar>=0.1.8